+++ src/core/component/const/cache.ts
@@ -7,7 +7,7 @@
  */
 
 import type { RenderObject } from 'core/component/render';
-import type { ComponentOptions as ComponentDriverOptions, ComponentDriver } from 'core/component/engines';
+import type { ComponentEngine, ComponentOptions as ComponentEngineOptions } from 'core/component/engines';
 import type { ComponentInterface, ComponentMeta, ComponentOptions } from 'core/component/interface';
 
 /**
@@ -18,7 +18,7 @@ export const componentParams = new Map<Function | string, ComponentOptions>();
 /**
  * Map of root components
  */
-export const rootComponents = Object.createDict<Promise<ComponentDriverOptions<ComponentDriver>>>();
+export const rootComponents = Object.createDict<Promise<ComponentEngineOptions<ComponentEngine>>>();
 
 /**
  * Link to an instance of the global root component