+++ src/core/component/engines/vue/vnode.ts
@@ -6,7 +6,7 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-import type { RenderContext, VNode } from 'vue';
+import Vue, { RenderContext, VNode } from 'vue';
 
 import { patchComponentVData } from 'core/component/vnode';
 import type { ComponentInterface } from 'core/component/interface';
@@ -23,11 +23,34 @@ export function cloneVNode(vnode: VNode): VNode {
  * Patches the specified VNode by using provided contexts
  *
  * @param vnode
- * @param ctx - component context
+ * @param component - component instance
  * @param renderCtx - render context
  */
-export function patchVNode(vnode: VNode, ctx: ComponentInterface, renderCtx: RenderContext): void {
+export function patchVNode(vnode: VNode, component: ComponentInterface, renderCtx: RenderContext): void {
 	patchComponentVData(vnode.data, renderCtx.data, {
-		patchAttrs: Boolean(ctx.unsafe.meta.params.inheritAttrs)
+		patchAttrs: Boolean(component.unsafe.meta.params.inheritAttrs)
 	});
+}
+
+/**
+ * Renders the specified VNode/s and returns the result
+ *
+ * @param vnode
+ * @param parent - parent component
+ */
+export function renderVNode(vnode: VNode, parent: ComponentInterface): Node;
+export function renderVNode(vnodes: VNode[], parent: ComponentInterface): Node[];
+export function renderVNode(vnode: CanArray<VNode>, parent: ComponentInterface): CanArray<Node> {
+	const vue = new Vue({
+		render: (c) => Object.isArray(vnode) ? c('div', vnode) : vnode
+	});
+
+	Object.set(vue, '$root', Object.create(parent.$root));
+	Object.set(vue, '$root.$remoteParent', parent);
+	Object.set(vue, '$root.unsafe', vue.$root);
+
+	const el = document.createElement('div');
+	vue.$mount(el);
+
+	return Object.isArray(vnode) ? Array.from(vue.$el.children) : vue.$el;
 }