+++ src/core/component/render-function/create-element.ts
@@ -18,19 +18,7 @@ import { createFakeCtx, initComponentVNode, FlyweightVNode } from 'core/componen
 import { applyDynamicAttrs } from 'core/component/render-function/v-attrs';
 import { registerComponent } from 'core/component/register';
 
-import {
-
-	supports,
-	cloneVNode,
-	minimalCtx,
-
-	CreateElement,
-
-	VNode,
-	VNodeData
-
-} from 'core/component/engines';
-
+import type { CreateElement, VNode, VNodeData } from 'core/component/engines';
 import type { FunctionalCtx, ComponentInterface, UnsafeComponentInterface } from 'core/component/interface';
 
 export const
@@ -50,23 +38,29 @@ export function wrapCreateElement(
 	baseCtx: ComponentInterface
 ): [CreateElement, Function[]] {
 	const
-		tasks = <Function[]>[];
+		tasks = <Function[]>[],
+		engine = baseCtx.$renderEngine;
+
+	const
+		{supports} = engine;
 
 	const wrappedCreateElement = <CreateElement>function wrappedCreateElement(
 		this: Nullable<ComponentInterface>,
 		tag: CanUndef<string>,
 		opts?: VNodeData,
 		children?: VNode[]
-	): VNode {
+	): CanPromise<VNode> {
 		// eslint-disable-next-line
 		'use strict';
 
 		const
 			ctx = this ?? baseCtx,
-			unsafe = <UnsafeComponentInterface><any>(ctx);
-
-		const
+			unsafe = <UnsafeComponentInterface><any>(ctx),
 			attrOpts = Object.isPlainObject(opts) ? opts.attrs : undefined;
+
+		const createElement = nativeCreateElement[$$.wrappedCreateElement] === true ?
+			nativeCreateElement :
+			nativeCreateElement.bind(ctx);
 
 		let
 			tagName = tag,
@@ -74,12 +68,12 @@ export function wrapCreateElement(
 
 		if (attrOpts == null) {
 			if (tag === 'v-render') {
-				return nativeCreateElement();
+				return createElement();
 			}
 
 		} else {
 			if (tag === 'v-render') {
-				return attrOpts.from ?? nativeCreateElement();
+				return attrOpts.from ?? createElement();
 			}
 
 			if (tagName?.[0] === '@') {
@@ -122,6 +116,8 @@ export function wrapCreateElement(
 			needLinkToEl = Boolean(flyweightComponent);
 
 		const needCreateFunctionalComponent =
+			!supports.regular ||
+
 			vnode == null &&
 			flyweightComponent == null &&
 			supports.functional &&
@@ -137,24 +133,27 @@ export function wrapCreateElement(
 				renderObject = c.componentTemplates[componentName];
 
 			if (renderObject == null) {
-				return nativeCreateElement();
+				return createElement();
 			}
 
 			const
-				node = nativeCreateElement('span', {...opts, tag: undefined}, children),
+				node = createElement('span', {...opts, tag: undefined}, children),
 				renderCtx = getComponentRenderCtxFromVNode(component, node, ctx);
 
 			let
 				baseCtx = <CanUndef<FunctionalCtx>>c.renderCtxCache[componentName];
 
 			if (baseCtx == null) {
-				baseCtx = Object.create(minimalCtx);
+				baseCtx = Object.create(engine.minimalCtx);
 
 				// @ts-ignore (access)
 				baseCtx.componentName = componentName;
 
 				// @ts-ignore (access)
 				baseCtx.meta = component;
+
+				// @ts-ignore (access)
+				component.params.functional = true;
 
 				// @ts-ignore (access)
 				baseCtx.instance = component.instance;
@@ -178,19 +177,25 @@ export function wrapCreateElement(
 				{initProps: true}
 			);
 
-			vnode = initComponentVNode(
+			const createComponentVNode = () => initComponentVNode(
 				execRenderObject(renderObject, fakeCtx),
 				fakeCtx,
 				renderCtx
 			);
+
+			if (supports.ssr && Object.isPromise(fakeCtx.unsafe.$initializer)) {
+				return fakeCtx.unsafe.$initializer.then(() => patchVNode(createComponentVNode()));
+			}
+
+			vnode = createComponentVNode();
 		}
 
 		if (vnode == null) {
-			// eslint-disable-next-line prefer-rest-params
-			vnode = nativeCreateElement.apply(unsafe, arguments);
+			vnode = createElement.apply(null, arguments);
 
+			// eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
 			if (vnode == null) {
-				return nativeCreateElement();
+				return createElement();
 			}
 
 			if (flyweightComponent != null) {
@@ -198,67 +203,92 @@ export function wrapCreateElement(
 			}
 		}
 
-		const
-			vData = vnode.data,
-			ref = vData && (vData[$$.ref] ?? vData.ref);
+		return patchVNode(vnode);
 
-		if (renderKey !== '') {
-			unsafe.renderTmp[renderKey] = cloneVNode(vnode);
-		}
+		function patchVNode<T extends VNode | FlyweightVNode>(vnode: T): T {
+			const
+				vData = vnode.data,
+				ref = vData != null && (vData[$$.ref] ?? vData.ref);
 
-		// Add $refs link if it doesn't exist
-		if (vData && ref != null && ctx !== baseCtx) {
-			vData[$$.ref] = ref;
-			vData.ref = `${ref}:${ctx.componentId}`;
-
-			Object.defineProperty(unsafe.$refs, ref, {
-				configurable: true,
-				enumerable: true,
-				get: () => {
-					const
-						r = baseCtx.unsafe.$refs,
-						l = r[`${ref}:${unsafe.$componentId}`] ?? r[`${ref}:${ctx.componentId}`];
-
-					if (l != null) {
-						return l;
-					}
-
-					if (vnode == null) {
-						return undefined;
-					}
-
-					return 'fakeInstance' in vnode ? vnode.fakeInstance : vnode.elm;
-				}
-			});
-		}
-
-		// Add $el link if it doesn't exist
-		if (needLinkToEl && 'fakeInstance' in vnode) {
-			Object.defineProperty(vnode.fakeInstance, '$el', {
-				enumerable: true,
-				configurable: true,
-
-				set(): void {
-					// Loopback
-				},
-
-				get(): CanUndef<Node> {
-					return vnode?.elm;
-				}
-			});
-		}
-
-		if (tasks.length > 0) {
-			for (let i = 0; i < tasks.length; i++) {
-				tasks[i](vnode);
+			if (renderKey !== '') {
+				unsafe.renderTmp[renderKey] = engine.cloneVNode(vnode);
 			}
 
-			tasks.splice(0);
-		}
+			// Add $refs link if it doesn't exist
+			if (vData != null && ref != null && ctx !== baseCtx) {
+				vData[$$.ref] = ref;
+				vData.ref = `${ref}:${ctx.componentId}`;
 
-		vnode.fakeContext = ctx;
-		return vnode;
+				Object.defineProperty(unsafe.$refs, ref, {
+					configurable: true,
+					enumerable: true,
+					get: () => {
+						const
+							r = baseCtx.unsafe.$refs,
+							l = r[`${ref}:${unsafe.$componentId}`] ?? r[`${ref}:${ctx.componentId}`];
+
+						if (l != null) {
+							return l;
+						}
+
+						return 'fakeInstance' in vnode ? (<FlyweightVNode>vnode).fakeInstance : vnode.elm;
+					}
+				});
+			}
+
+			// Add $el link if it doesn't exist
+			if (needLinkToEl && 'fakeInstance' in vnode) {
+				Object.defineProperty((<FlyweightVNode>vnode).fakeInstance, '$el', {
+					enumerable: true,
+					configurable: true,
+
+					set(): void {
+						// Loopback
+					},
+
+					get(): CanUndef<Node> {
+						return vnode.elm;
+					}
+				});
+			}
+
+			if (tasks.length > 0) {
+				for (let i = 0; i < tasks.length; i++) {
+					tasks[i](vnode);
+				}
+
+				tasks.splice(0);
+			}
+
+			vnode.fakeContext = ctx;
+			return vnode;
+		}
 	};
+
+	wrappedCreateElement[$$.wrappedCreateElement] = true;
+
+	if (supports.ssr) {
+		const wrappedAsyncCreateElement = <CreateElement>function wrappedAsyncCreateElement(
+			this: Nullable<ComponentInterface>,
+			tag: CanUndef<string>,
+			opts?: VNodeData,
+			children?: VNode[]
+		): CanPromise<VNode> {
+			if (children != null && children.length > 0) {
+				children = children.flat();
+
+				// eslint-disable-next-line @typescript-eslint/unbound-method
+				if (children.some(Object.isPromise)) {
+					return Promise.all(children).then((children) => wrappedCreateElement.call(this, tag, opts, children));
+				}
+			}
+
+			// eslint-disable-next-line prefer-rest-params
+			return wrappedCreateElement.apply(this, arguments);
+		};
+
+		return [wrappedAsyncCreateElement, tasks];
+	}
 
 	return [wrappedCreateElement, tasks];
 }