+++ src/super/i-block/i-block.ts
@@ -60,6 +60,7 @@ import {
 
 } from 'core/component';
 
+import remoteState from 'core/component/state';
 import * as init from 'core/component/construct';
 
 import 'super/i-block/directives';
@@ -540,7 +541,7 @@ export default abstract class iBlock extends ComponentInterface {
 	 * Link to i18n function, that will be used to localize of string literals
 	 */
 	@prop(Function)
-	readonly i18n: typeof i18n = defaultI18n;
+	readonly i18n: typeof i18n = ((i18n));
 
 	/**
 	 * Link to a remote state object.
@@ -549,9 +550,9 @@ export default abstract class iBlock extends ComponentInterface {
 	 * that can't be initialized within a component directly. You can modify this object outside from components,
 	 * but remember, that these mutations may force re-render of all components.
 	 */
-	@computed({watchable: true, dependencies: ['r.remoteState']})
-	get remoteState(): this['r']['remoteState'] {
-		return this.r.remoteState;
+	@computed({watchable: true})
+	get remoteState(): typeof remoteState {
+		return remoteState;
 	}
 
 	/**
@@ -722,6 +723,14 @@ export default abstract class iBlock extends ComponentInterface {
 	@computed({replace: false})
 	get isNotRegular(): boolean {
 		return Boolean(this.isFunctional || this.isFlyweight);
+	}
+
+	/**
+	 * True if the current component is rendered by using server-side rendering
+	 */
+	@computed({replace: false})
+	get isSSR(): boolean {
+		return this.$renderEngine.supports.ssr;
 	}
 
 	/**
@@ -1576,7 +1585,7 @@ export default abstract class iBlock extends ComponentInterface {
 		optsOrHandler: AsyncWatchOptions | RawWatchHandler<this, T>,
 		handlerOrOpts?: RawWatchHandler<this, T> | AsyncWatchOptions
 	): void {
-		if (this.isFlyweight) {
+		if (this.isFlyweight || this.isSSR) {
 			return;
 		}
 
@@ -2016,16 +2025,21 @@ export default abstract class iBlock extends ComponentInterface {
 				this.state.initFromStorage() || []
 			);
 
-			if (this.isNotRegular || this.dontWaitRemoteProviders) {
+			if (
+				(this.isNotRegular || this.dontWaitRemoteProviders) &&
+				!this.$renderEngine.supports.ssr
+			) {
 				if (tasks.length > 0) {
-					return $a.promise(SyncPromise.all(tasks), label).then(done, doneOnError);
+					const res = $a.promise(SyncPromise.all(tasks), label).then(done, doneOnError);
+					this.$initializer = res;
+					return res;
 				}
 
 				done();
 				return;
 			}
 
-			return this.nextTick(label).then((() => {
+			const res = this.nextTick(label).then((() => {
 				const
 					{$children: childComponents} = this;
 
@@ -2079,6 +2093,9 @@ export default abstract class iBlock extends ComponentInterface {
 
 				return $a.promise(SyncPromise.all(tasks), label).then(done, doneOnError);
 			}));
+
+			this.$initializer = res;
+			return res;
 
 		} catch (err) {
 			doneOnError(err);
@@ -2527,7 +2544,7 @@ export default abstract class iBlock extends ComponentInterface {
 
 	/** @override */
 	protected onCreatedHook(): void {
-		if (this.isFlyweight) {
+		if (this.isFlyweight || this.isSSR) {
 			this.componentStatusStore = 'ready';
 			this.isReadyOnce = true;
 		}
@@ -2599,9 +2616,4 @@ export default abstract class iBlock extends ComponentInterface {
 			delete classesCache.dict.els?.[this.componentId];
 		} catch {}
 	}
-}
-
-function defaultI18n(this: iBlock, ...args: unknown[]): string {
-	// eslint-disable-next-line @typescript-eslint/no-extra-parens
-	return (Object.isFunction(this.r.i18n) ? this.r.i18n : ((i18n))).apply(this.r, args);
 }