+++ src/core/component/interface/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.??? (2021-??-??)
+
+#### :rocket: New Feature
+
+* Added a new property `$initializer`
+* Added a new property `$renderEngine`
+
 ## v3.0.0-rc.126 (2021-01-26)
 
 #### :boom: Breaking Change